+++ src/lang/program_util.cpp
@@ -50,7 +50,10 @@ bool ProgramUtil::isNop(const Operation &op) {
              op.source.value == Number::ZERO &&
              (op.type == Operation::Type::ADD ||
               op.type == Operation::Type::SUB ||
-              op.type == Operation::Type::CLR)) {
+              op.type == Operation::Type::CLR ||
+              op.type == Operation::Type::FIL ||
+              op.type == Operation::Type::ROL ||
+              op.type == Operation::Type::ROR)) {
     return true;
   } else if (op.source.type == Operand::Type::CONSTANT &&
              op.source.value == Number::ONE &&
@@ -59,7 +62,9 @@ bool ProgramUtil::isNop(const Operation &op) {
                op.type == Operation::Type::DIF ||
                op.type == Operation::Type::DIR ||
                op.type == Operation::Type::POW ||
-               op.type == Operation::Type::BIN))) {
+               op.type == Operation::Type::BIN ||
+               op.type == Operation::Type::ROL ||
+               op.type == Operation::Type::ROR))) {
     return true;
   }
   return false;
@@ -98,8 +103,9 @@ size_t ProgramUtil::numOps(const Program &p, Operand::Type type) {
 bool ProgramUtil::isArithmetic(Operation::Type t) {
   return (t != Operation::Type::NOP && t != Operation::Type::DBG &&
           t != Operation::Type::LPB && t != Operation::Type::LPE &&
-          t != Operation::Type::CLR && t != Operation::Type::SEQ &&
-          t != Operation::Type::PRG);
+          t != Operation::Type::CLR && t != Operation::Type::FIL &&
+          t != Operation::Type::ROL && t != Operation::Type::ROR &&
+          t != Operation::Type::SEQ && t != Operation::Type::PRG);
 }
 
 bool ProgramUtil::isCommutative(Operation::Type t) {
@@ -165,7 +171,9 @@ bool ProgramUtil::isReadingCell(const Operation &op, int64_t cell) {
 }
 
 bool ProgramUtil::isWritingRegion(Operation::Type t) {
-  return (t == Operation::Type::CLR || t == Operation::Type::PRG);
+  return (t == Operation::Type::CLR || t == Operation::Type::FIL ||
+          t == Operation::Type::ROL || t == Operation::Type::ROR ||
+          t == Operation::Type::PRG);
 }
 
 bool ProgramUtil::hasRegionOperation(const Program &p) {
@@ -216,42 +224,84 @@ bool ProgramUtil::areIndependent(const Operation &op1, const Operation &op2) {
   return true;
 }
 
+void collectUsedMemoryCells(const Operand &operand, int64_t region_length,
+                            std::unordered_set<int64_t> *used_cells,
+                            int64_t &largest_used) {
+  if (operand.type != Operand::Type::DIRECT) {
+    return;
+  }
+  int64_t start, end;
+  int64_t base = operand.value.asInt();
+  if (region_length >= 0) {
+    start = base;
+    end = base + region_length;
+  } else {
+    start = std::max<int64_t>(base + region_length + 1, 0);
+    end = base + 1;
+  }
+  for (int64_t cell = start; cell < end; cell++) {
+    if (used_cells) {
+      used_cells->insert(cell);
+    }
+    largest_used = std::max(largest_used, cell);
+  }
+}
+
 bool ProgramUtil::getUsedMemoryCells(const Program &p,
-                                     std::unordered_set<int64_t> &used_cells,
+                                     std::map<UID, Program> *prg_refs,
+                                     std::unordered_set<int64_t> *used_cells,
                                      int64_t &largest_used,
                                      int64_t max_memory) {
+  largest_used = 0;
   for (const auto &op : p.ops) {
     int64_t region_length = 1;
     if (op.source.type == Operand::Type::INDIRECT ||
-        op.target.type == Operand::Type::INDIRECT ||
-        op.type == Operation::Type::PRG) {
+        op.target.type == Operand::Type::INDIRECT) {
       return false;
     }
-    if (op.type == Operation::Type::LPB || op.type == Operation::Type::CLR) {
+    if (op.type == Operation::Type::PRG) {
+      if (op.source.type != Operand::Type::CONSTANT) {
+        return false;
+      }
+      auto sub_uid = UID::castFromInt(op.source.value.asInt());
+      if (!prg_refs || prg_refs->find(sub_uid) == prg_refs->end()) {
+        return false;
+      }
+      const auto &sub = prg_refs->at(sub_uid);
+      region_length = std::max<int64_t>(sub.getDirective("inputs"),
+                                        sub.getDirective("outputs"));
+    } else if (op.type == Operation::Type::LPB ||
+               op.type == Operation::Type::CLR ||
+               op.type == Operation::Type::FIL ||
+               op.type == Operation::Type::ROL ||
+               op.type == Operation::Type::ROR) {
       if (op.source.type == Operand::Type::CONSTANT) {
         region_length = op.source.value.asInt();
       } else {
         return false;
       }
     }
-    if (region_length > max_memory && max_memory >= 0) {
+    if (std::abs(region_length) > max_memory && max_memory >= 0) {
       return false;
     }
+    collectUsedMemoryCells(op.source, region_length, used_cells, largest_used);
+    collectUsedMemoryCells(op.target, region_length, used_cells, largest_used);
+  }
+  return true;
+}
+
+int64_t ProgramUtil::getLargestDirectMemoryCellWithoutRegions(
+    const Program &p) {
+  int64_t largest = 0;
+  for (const auto &op : p.ops) {
     if (op.source.type == Operand::Type::DIRECT) {
-      for (int64_t i = 0; i < region_length; i++) {
-        used_cells.insert(op.source.value.asInt() + i);
-      }
+      largest = std::max<int64_t>(largest, op.source.value.asInt());
     }
     if (op.target.type == Operand::Type::DIRECT) {
-      for (int64_t i = 0; i < region_length; i++) {
-        used_cells.insert(op.target.value.asInt() + i);
-      }
+      largest = std::max<int64_t>(largest, op.target.value.asInt());
     }
   }
-  const auto max = [](int64_t a, int64_t b) { return std::max(a, b); };
-  largest_used =
-      std::accumulate(used_cells.begin(), used_cells.end(), (int64_t)0, max);
-  return true;
+  return largest;
 }
 
 bool ProgramUtil::getUsedUninitializedCells(const Program &p,
@@ -284,7 +334,8 @@ bool ProgramUtil::getUsedUninitializedCells(const Program &p,
       }
     }
     // check region operations
-    if (op.type == Operation::Type::CLR) {
+    if (op.type == Operation::Type::CLR || op.type == Operation::Type::FIL ||
+        op.type == Operation::Type::ROL || op.type == Operation::Type::ROR) {
       if (op.source.type == Operand::Type::CONSTANT) {
         // start of region (direct memory cell)
         const auto t = op.target.value.asInt();
@@ -303,19 +354,6 @@ bool ProgramUtil::getUsedUninitializedCells(const Program &p,
     }
   }
   return true;
-}
-
-int64_t ProgramUtil::getLargestDirectMemoryCell(const Program &p) {
-  int64_t largest = 0;
-  for (const auto &op : p.ops) {
-    if (op.source.type == Operand::Type::DIRECT) {
-      largest = std::max<int64_t>(largest, op.source.value.asInt());
-    }
-    if (op.target.type == Operand::Type::DIRECT) {
-      largest = std::max<int64_t>(largest, op.target.value.asInt());
-    }
-  }
-  return largest;
 }
 
 bool ProgramUtil::swapDirectOperandCells(Program &p, int64_t cell1,